pyproject.toml
[tool.poetry]
name = "botaclan"
version = "0.1.0"
description = "Bataclan's official bot"
license = "Apache-2.0"
authors = [
    "Gabriel Tiossi <gabrieltiossi@gmail.com>"
]
maintainers = [
]
readme = "README.md"
homepage = ""
repository = "https://github.com/bataclanofficial/botaclan"
documentation = ""
keywords = [
    "discord",
    "bataclan",
    "python",
    "bot"
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3.8",
    "Topic :: Communications :: Chat"
]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.3.3"}
coloredlogs = "^14.0"
vyper-config = "^0.6.2"
google-api-python-client = "^1.8.3"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
dateparser = "^0.7.4"
sentry-sdk = "^0.14.4"
importlib-metadata = "^1.6.0"
fastapi = "^0.54.2"
uvicorn = "^0.11.5"
youtube_dl = "^2020.5.8"
django = "^3.0.6"
humanfriendly = "^8.2"
soundcloud-lib = "^0.5.0"
pyspotify = "^2.1.3"
spotipy = "^2.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
black = "^19.10b0"
flake8 = "^3.7.9"

[tool.poetry.scripts]
botaclan = "botaclan.__main__"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
